requirements.txt
flask
ultralytics
Pillow
torch
opencv-python-headless
pandas
requests
